tested possibilities and updated plan
Did a few tests inside of main.py to understand the project further and modify my plans in plan.txt.

In main.py, you'll find two sections that were the two main tests. Both have explanations of which parts of them are viable or not, and what the final solution may be.